Refactor client list
